chore(build): tidy up file structure and add grunt file
The grunt file is not really needed but it demonstrates how easy it is to do a jshint and testacular run
Also, this grunt build allows us to run both jshint and testacular in the .travis CI build.
Currently the demo page is self contained but it would be good to think about breaking it out into a partial that could be used by the super angular-ui component.