fix(setViewValue): only update the model with the real date on blur
Getting the timing right on updating the model is really tricky.
Now, if you enter text in the input box directly it updates the model
as for any input box.  But when you blur away, it uses the date picker
to convert the value to a date object.

Previously, as soon as you made changes in the input box it was
immediately updating the date, which was preventing typing.

Closes #18